requirements.txt
asgiref==3.5.2
bootstrap4==0.1.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.0.0
Django==4.0.6
django-cors-headers==3.13.0
django-crispy-bootstrap==0.1.1.1
django-crispy-forms==1.14.0
django-import-export==3.0.1
et-xmlfile==1.1.0
MarkupPy==1.14
mysqlclient==2.1.1
numpy==1.23.5
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.5.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
tzdata==2022.7
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
